+++ src/portfolio.jl
@@ -1,0 +1,118 @@
+module Portfolio
+
+using JuMP, Ipopt, Statistics, LinearAlgebra
+
+import ..OperationsResearchModels: solve
+
+export PortfolioProblem, PortfolioResult
+
+
+
+
+"""
+    PortfolioProblem
+
+# Description
+
+Defines the portfolio optimization problem.
+
+# Fields
+
+- `returns::Matrix{<:Real}`: A matrix of historical returns for the assets.
+- `thresholdreturn::Real`: The minimum expected return required for the portfolio.
+
+"""
+struct PortfolioProblem
+    returns::Matrix{<:Real}
+    thresholdreturn::Real 
+end 
+
+
+
+"""
+    PortfolioResult
+
+# Description
+
+A structure to hold the result of the portfolio optimization problem.
+
+# Fields
+
+- `weights::Vector{Float64}`: The optimal weights for the assets in the portfolio.
+- `expectedreturn::Float64`: The expected return of the optimal portfolio.
+- `model::JuMP.Model`: The JuMP model used to solve the problem.
+"""
+struct PortfolioResult
+    weights::Vector{Float64}
+    expectedreturn::Float64
+    model::JuMP.Model
+end
+
+
+
+"""
+    solve(problem)
+
+# Description
+
+Solves a portfolio optimization problem given by an object of in type `PortfolioProblem`.
+The optimization problem is formulated as a quadratic programming problem where the objective 
+is to minimize the portfolio variance (risk) subject to constraints on the expected return and the weights.
+
+Mathematically, the problem can be stated as:
+
+Minimize: w' * Σ * w
+Subject to:
+- sum(w) == 1 (the weights must sum to 1)
+- sum(w[i] * μ[i] for i in 1:m) >= thresholdreturn
+- 0 <= w[i] <= 1 for all i (weights must be between 0 and 1)
+
+Where:
+- w is the vector of asset weights
+- Σ is the covariance matrix of asset returns
+- μ is the vector of expected returns for each asset
+- thresholdreturn is the minimum expected return required for the portfolio
+
+# Arguments
+
+- `problem::PortfolioProblem`: The problem in type of PortfolioProblem
+
+# Returns
+
+- `PortfolioResult`: The result of the portfolio optimization problem, containing the optimal weights, expected return, and the JuMP model used to solve the problem.
+
+"""
+function solve(p::PortfolioProblem)::PortfolioResult
+    
+    model = Model(Ipopt.Optimizer)
+
+    MOI.set(model, MOI.Silent(), true)
+
+    means = mean(p.returns, dims = 1)
+
+    covmat = cov(p.returns)
+
+    _, m = size(p.returns)
+
+    @variable(model, 0 <= w[1:m] <= 1)
+    
+    @constraint(model, sum(w) == 1)
+
+    @constraint(model, sum(w[i] * means[i] for i in 1:m) >= p.thresholdreturn)
+
+    @objective(model, Min, w' * covmat * w)
+
+    optimize!(model)
+
+    weights = value.(w)
+
+    expectedreturn = sum(weights[i] * means[i] for i in 1:m)
+
+    return PortfolioResult(weights, expectedreturn, model)
+
+end 
+
+
+
+
+end # end of module